agent: support FunctionTool handling in LLMAgent (#29)
* agent: support FunctionTool handling in LLMAgent

During post-processing, if call the requested function tool
and generate the function response event. The event is appended
in the session's events. In the next iteration of baseFlow's loop,
the response is sent to LLM.

The control flow and function names correspond to Python/Java's.

Added a simple test using gemini 2.0 flash.